IBL, Bloom, Shadows and SSAO

+++ Source/Models/Material.h
@@ -23,9 +23,9 @@ namespace Models
 {
 	struct Material
     {
-	    usize     albedo          = 0;
-	    usize     normal          = 0;
-	    usize     aoRghMtl        = 0;
+	    u32       albedo          = 0;
+	    u32       normal          = 0;
+	    u32       aoRghMtl        = 0;
 		glm::vec4 albedoFactor    = {1.0f, 1.0f, 1.0f, 1.0f};
 		f32       roughnessFactor = 1.0f;
 		f32       metallicFactor  = 1.0f;
